pyproject.toml
[project]
name = "sharatiy-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langdetect>=1.0.9",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
